Fix SSL certificate issue
This commit fixes an issue related with ssl certificate.

If the host machine did not have an SSL certificate, the CURL would not be able to not connect to the destiny.
Now, at least with debug turn on, the SSL certification host validation will be ignored and the CURL will be able to connect with the URL destination.